% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/write_mult.R
\name{write_mult}
\alias{write_mult}
\title{Write multiple files from one table using column name as new file name.}
\usage{
write_mult(x, type = "csv")
}
\arguments{
\item{x}{table of data with column headers.}

\item{type}{the type of file to be made (csv, txt).}
}
\value{
The created files saved to current working directory.
}
\description{
Write multiple files from one table using column name as new file name.
}
\examples{
write_mult(iris, txt)
write_mult(iris)
}

